fix(cli): property priority
Introduced a merge function which is in charge to merge all properties, giving priority to single declaration property options

Closes #2538
